+++ Configuration/Icons.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    'tx-faqseo-svgicon' => [
+        'provider' => \TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
+        'source' => 'EXT:faq_seo/Resources/Public/Icons/Extension.svg',
+    ],
+];