+++ CHANGES.md
@@ -1,3 +1,7 @@
+## Version 1.1.0
+- [FEATURE] Add FaqItem Model
+- [WIP] Add Import Backend module
+
 ## Version 1.0.0
 - [TASK] Add FAQ Styling
 - [TASK] Add JS Function if bootstrap not installed